refactor PADIM implementation
