+++ anomalib/models/padim/config_fldr.yaml
@@ -1,0 +1,97 @@
+dataset:
+  name: Test
+  format: folder
+  path: /home/djameln/Desktop/bottle_folder
+  normal: good # name of the folder containing normal images.
+  abnormal: broken_large # name of the folder containing abnormal images.
+  #  task: classification # classification or segmentation
+  mask: null #optional
+  extensions: null
+  split_ratio: 0.2 # ratio of the normal images that will be used to create a test split
+  seed: 0
+  image_size: 256
+  train_batch_size: 32
+  test_batch_size: 32
+  num_workers: 0
+  transform_config: null
+  create_validation_set: true
+  tiling:
+    apply: false
+    tile_size: null
+    stride: null
+    remove_border_count: 0
+    use_random_tiling: False
+    random_tile_count: 16
+
+model:
+  name: padim
+  backbone: resnet18
+  layers:
+    - layer1
+    - layer2
+    - layer3
+  metric: auc
+  normalization_method: min_max # options: [none, min_max, cdf]
+  threshold:
+    image_default: 3
+    pixel_default: 3
+    adaptive: true
+
+project:
+  seed: 42
+  path: ./results
+  log_images_to: ["local"]
+  logger: false
+  save_to_csv: false
+
+optimization:
+  openvino:
+    apply: false
+
+# PL Trainer Args. Don't add extra parameter here.
+trainer:
+  accelerator: null
+  accumulate_grad_batches: 1
+  amp_backend: native
+  auto_lr_find: false
+  auto_scale_batch_size: false
+  auto_select_gpus: false
+  benchmark: false
+  check_val_every_n_epoch: 1 # Don't validate before extracting features.
+  checkpoint_callback: true
+  default_root_dir: null
+  deterministic: false
+  fast_dev_run: false
+  gpus: 1
+  gradient_clip_val: 0
+  limit_predict_batches: 1.0
+  limit_test_batches: 1.0
+  limit_train_batches: 1.0
+  limit_val_batches: 1.0
+  log_every_n_steps: 50
+  log_gpu_memory: null
+  max_epochs: 1
+  max_steps: -1
+  min_epochs: null
+  min_steps: null
+  move_metrics_to_cpu: false
+  multiple_trainloader_mode: max_size_cycle
+  num_nodes: 1
+  num_processes: 1
+  num_sanity_val_steps: 0
+  overfit_batches: 0.0
+  plugins: null
+  precision: 32
+  prepare_data_per_node: true
+  process_position: 0
+  profiler: null
+  progress_bar_refresh_rate: null
+  replace_sampler_ddp: true
+  stochastic_weight_avg: false
+  sync_batchnorm: false
+  terminate_on_nan: false
+  tpu_cores: null
+  track_grad_norm: -1
+  val_check_interval: 1.0 # Don't validate before extracting features.
+  weights_save_path: null
+  weights_summary: top
